Effectuer un retour

+++ src/main/java/org/miage/trainprojet/Control/ReservationAssembler.java
@@ -9,6 +9,7 @@ import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.server.RepresentationModelAssembler;
 import org.springframework.stereotype.Component;
 
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.stream.StreamSupport;
 
@@ -17,29 +18,53 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @Component
 public class ReservationAssembler implements RepresentationModelAssembler<Reservation,EntityModel<Reservation>> {
+
     @Override
     public EntityModel<Reservation> toModel(Reservation reservation) {
-        return EntityModel.of(reservation,
+        DateTimeFormatter formatters = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
+        EntityModel<Reservation> link = EntityModel.of(reservation,
                 linkTo(methodOn(ReservationRepresentation.class)
                         .getOneReservation(reservation.getId())).withSelfRel(),
                 linkTo(methodOn(TrajetRepresentation.class)
-                        .rechercheRetour(reservation.getAller().getArrivee(), reservation.getAller().getDepart(),reservation.getAller().getJour().plusDays(1).toString(),2)).withRel("Rechercher un retour"),
-                linkTo(methodOn(ReservationRepresentation.class)
-                        .patchRetour(reservation.getId(),null)).withRel("Ajouter un retour"),
-                linkTo(methodOn(ReservationRepresentation.class)
-                        .patchConfirme(reservation.getId())).withRel("Confirmer cette reservation"),
-                linkTo(methodOn(ReservationRepresentation.class)
-                        .deleteReservation(reservation.getId())).withRel("Annuler cette réservation"));
+                        .getOneTrajet(reservation.getAller().getId())).withRel("Trajet aller"));
+
+        if(reservation.getRetour() == null){
+            link.add(linkTo(methodOn(TrajetRepresentation.class)
+                    .rechercheRetour(reservation.getId(), reservation.getAller().getArrivee(), reservation.getAller().getDepart(),reservation.getAller().getJour().plusDays(1).format(formatters),2)).withRel("Rechercher un retour"));
+        } else {
+            link.add(linkTo(methodOn(TrajetRepresentation.class)
+                    .getOneTrajet(reservation.getAller().getId())).withRel("Trajet retour"));
+        }
+
+        if(!reservation.isConfirme()){
+            link.add(linkTo(methodOn(ReservationRepresentation.class)
+                    .patchConfirme(reservation.getId())).withRel("Confirmer cette reservation"));
+
+            link.add(linkTo(methodOn(ReservationRepresentation.class)
+                    .deleteReservation(reservation.getId())).withRel("Annuler cette réservation"));
+        }
+
+        return link;
     }
 
     @Override
     public CollectionModel<EntityModel<Reservation>> toCollectionModel(Iterable<? extends Reservation> entities) {
-        List<EntityModel<Reservation>> reservatiobModel = StreamSupport
+        List<EntityModel<Reservation>> reservationModel = StreamSupport
                 .stream(entities.spliterator(), false)
                 .map(i-> toModel(i))
                 .toList();
 
-        return CollectionModel.of(reservatiobModel,
+        return CollectionModel.of(reservationModel,
+                linkTo(methodOn(TrajetRepresentation.class).getAllTrajets()).withSelfRel());
+    }
+
+    public CollectionModel<EntityModel<Reservation>> toCollectionModelRetour(Iterable<? extends Reservation> entities) {
+        List<EntityModel<Reservation>> reservationModel = StreamSupport
+                .stream(entities.spliterator(), false)
+                .map(i-> toModel(i))
+                .toList();
+
+        return CollectionModel.of(reservationModel,
                 linkTo(methodOn(TrajetRepresentation.class).getAllTrajets()).withSelfRel());
     }
 }
